+++ scripts/twenty-one/rulesets/index.js
@@ -1,0 +1,2 @@
+export { submitGoblinRoll } from "./goblin.js";
+export { submitStandardRoll } from "./standard.js";